services/admin-service/requirements.txt
fastapi>=0.109.1
uvicorn[standard]>=0.24.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0
redis>=5.0.7
celery>=5.3.0
python-jose[cryptography]>=3.3.4
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.8
httpx>=0.25.2
aiohttp>=3.10.11
beautifulsoup4>=4.12.0
lxml>=4.9.0
psutil>=5.9.0
cryptography>=42.0.0
structlog>=23.2.0
python-dotenv>=1.0.0